# Conflicts:
#	Source/Backends/DX12/Layer/Source/Compiler/DXIL/Blocks/DXILPhysicalBlockFunction.cpp
#	Source/UIX/Studio/App/DesktopApp.axaml.cs

+++ Source/Libraries/Backend/Generator/Source/ShaderExportGenerator.cpp
@@ -36,7 +36,7 @@ bool ShaderExportGenerator::Generate(Schema &schema, Language language, SchemaSt
         if (!message.attributes.GetBool("no-sguid")) {
             Field& sguid = *message.fields.emplace(message.fields.begin());
             sguid.name = "sguid";
-            sguid.type = "uint16";
+            sguid.type = "uint32";
 
             // Attributes
             sguid.attributes.Add("bits", std::to_string(kShaderSGUIDBitCount));
